// src/types/schema.ts
export interface components {
    schemas: {
// ...
        EnumEventType: "ACADEMIC" | "SOCIAL" | "SPORTS" | "CAREER" | "VOLUNTEERING" | "WORKSHOP";
        /**
         * @description Clasificarea locației pentru filtrare rapidă.
         * @enum {string}
         */
        EnumLocationType: "IN_CAMPUS" | "IN_CITY" | "OUTSIDE_CITY" | "ONLINE";
// ...
        EnumOrganizerType: "STUDENT_ORG" | "UNIVERSITY_DEPT" | "COMPANY" | "OTHER";
// ...
        AssociationSimple: {
            /** Format: uuid */
            id?: string;
            name?: string;
            logoUrl?: string;
            type?: components["schemas"]["EnumOrganizerType"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}